requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.6.3
bidict==0.23.1
blinker==1.9.0
cachetools==4.2.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
eventlet==0.39.1
Flask==3.1.0
Flask-SocketIO==5.3.6
greenlet==3.2.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
instagrapi==2.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==25.0
pdfminer.six==20250327
pdfplumber==0.11.6
pefile==2023.2.7
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pypdfium2==4.30.1
pyperclip==1.9.0
PySocks==1.7.1
python-dotenv==1.1.0
python-engineio==4.9.1
python-socketio==5.11.2
python-telegram-bot==20.0
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
setuptools==80.0.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
stripe==12.0.1
tornado==6.1
tweepy==4.15.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==1.26.18
Werkzeug==3.1.3
wsproto==1.2.0
